Send -march to gcc
